requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
tk>=0.1.0
